test: move pipe_backgroundping_race_test to pipe_test
